[FLINK-34455] Move RestoreMode from flink-runtime to flink-core
