+++ flink-runtime/src/main/java/org/apache/flink/runtime/state/SharedStateRegistry.java
@@ -18,8 +18,8 @@
 
 package org.apache.flink.runtime.state;
 
+import org.apache.flink.core.execution.RestoreMode;
 import org.apache.flink.runtime.checkpoint.CompletedCheckpoint;
-import org.apache.flink.runtime.jobgraph.RestoreMode;
 
 import java.util.Set;
